+++ MuscleTrainingRecords00/MuscleTrainingRecords00/RecordsModel.cs
@@ -16,7 +16,7 @@ namespace MuscleTrainingRecords00
 
         public string M_name { get; set; } //筋トレ名前
 
-        public int M_weight { get; set; } //重量
+        public double M_weight { get; set; } //重量
 
         public int M_leg { get; set; } //回数
 
@@ -25,13 +25,8 @@ namespace MuscleTrainingRecords00
         public string M_date { get; set; } //日付
 
 
-
-
-        //[ForeignKey(typeof(SettingModel))]
-        // public int Set_no { get; set; } //Setting表の外部キー
-
         /********************インサートメソッド**********************/
-        public static void InsertRecords(int m_weight, int m_leg, int m_set, string m_date)
+        public static void InsertRecords(double m_weight, int m_leg, int m_set, string m_date)
         {
             //データベースに接続する
             using (SQLiteConnection db = new SQLiteConnection(App.dbPath))
@@ -53,7 +48,7 @@ namespace MuscleTrainingRecords00
         }
 
         /********************インサートメソッド RecordListPage　追加**********************/
-        public static void InsertRe(int m_no, string m_name, int m_weight, int m_leg, int m_set, string m_date)
+        public static void InsertRe(int m_no, string m_name, double m_weight, int m_leg, int m_set, string m_date)
         {
             //データベースに接続する
             using (SQLiteConnection db = new SQLiteConnection(App.dbPath))
@@ -147,7 +142,7 @@ namespace MuscleTrainingRecords00
         }
 
         /********************アップデートメソッド RecordPageに使用**************************************/
-        public static void UpdateRecord(int m_no, int m_weight, int m_leg, int m_set, string m_date)
+        public static void UpdateRecord(int m_no, double m_weight, int m_leg, int m_set, string m_date)
         {
             using (SQLiteConnection db = new SQLiteConnection(App.dbPath))
             {
